backend/requirements.txt
fastapi>=0.110.0,<0.116.0
uvicorn[standard]>=0.23.0,<0.32.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
SQLAlchemy>=2.0.25,<3.0.0
httpx>=0.25.0,<0.28.0
python-multipart>=0.0.6,<0.0.10
PyJWT>=2.6.0,<3.0.0
redis>=5.0.0,<6.0.0
apscheduler>=3.10.4,<4.0.0
pypdf>=4.2.0,<5.0.0
python-docx>=1.1.0,<1.2.0
orjson>=3.9.0,<4.0.0
